# 8311594: Avoid GlobalSession liveness check