@mcimadamore 2023-07-31
> For liveness check, a solution on the top of my head is to push the branch to the slow path. Instead of having one `state` field, we can have 2, which can take values according to the state of the session, (0, x): Open, (1, 0): Closing, (1, 1): Closed. These can be marked as stable, since they do not transit from value 1 to value 0. Then, we can special case the global session to have the `openState` field being 1, and we do type branching when encountering a 1 when reading this state. This have the advantage that it does not affect the execution of other sessions, as well as being able to fold the check for the global ones.
> 
> For ownership check, maybe we can tell the compiler to trust this particular field. Stable does not work here, though maybe we can add a dummy `Thread` object to use as the pivot value instead of `null`. But comparing to another pointer is a bit tad more expensive than comparing to `null`, so changing the compiler seems more preferable.
> 
> For range check, maybe we can do the same as liveness one, instead of failing on out of bounds, we branch on a constant field to see if it is a global one, and only throw if it is not.
> 
> Thanks a lot.

What you suggest are sensible optimization tactics. However, it is unclear how much this path needs to be optimized further. In most cases, when we see a significant performance difference between MS and Unsafe, the liveness check is never the culprit, the bound check is.

And, when using tactics to e.g. generate wrapper segments on the fly to avoid bound checks [1], we can observe that liveness check is also gone from the corresponding assembly. So, this optimization seems to be in YAGNI territory, at least for now.

[1] - https://mail.openjdk.org/pipermail/panama-dev/2023-July/019487.html